refactor #21 - String Extension 분리
